requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
Django==5.0.7
django-crispy-forms==2.3
python-nmap==0.7.1
sqlparse==0.5.1
tzdata==2024.1
geoip2
djangorestframework


